Linux 2.2.26
o   CAN-2004-0077: behave safely in case of do_munmap()   (Solar Designer)
     failures in mremap(2)
o   CAN-2003-0984: /dev/rtc can leak parts of kernel      (Solar Designer)
     memory to unprivileged users (2.4 backport)
o   CAN-2003-0244: hashing exploits in network stack      (David S. Miller)
o   update_atime() performance improvement (2.4 backport) (Solar Designer)
o   ability to swapoff after a device file might          (Solar Designer)
     have been re-created
o   MAINTAINERS correction for Kernel 2.2 and 2.2 fixes   (Marc-Christian Petersen)
o   fixed some typos                                      (Solar Designer,
                                                           Marc-Christian Petersen)

+++ drivers/sgi/char/graphics.c
@@ -105,7 +105,7 @@ sgi_graphics_ioctl (struct inode *inode, struct file *file, unsigned int cmd, un
 
 		if (board >= boards)
 			return -EINVAL;
-		if (max_len < sizeof (struct gfx_getboardinfo_args))
+		if (max_len < (int)sizeof (struct gfx_getboardinfo_args))
 			return -EINVAL;
 		if (max_len > cards [board].g_board_info_len)
 			max_len = cards [boards].g_board_info_len;
